Separate error variants from app error class definition
